backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-dotenv
pydantic
pydantic-settings
email-validator
python-multipart
requests
ruff
